feat: refactor the useCheckout hook
